Use a precompiled NIF
This commit introduces the ability to download a precompiled NIF file
and skip the need to have Rust toolchain installed.

It is going to work for the majority of popular Operational Systems. For
those that are not supported, the user can still force compilation by
setting the "HTML5EVER_BUILD" environment variable to "true" or "1".
The following targets are supported:

    aarch64-apple-darwin - MacOS on M1
    x86_64-apple-darwin
    x86_64-unknown-linux-gnu
    x86_64-unknown-linux-musl
    arm-unknown-linux-gnueabihf
    x86_64-pc-windows-gnu
    x86_64-pc-windows-msvc

The supported NIF versions are "2.14", "2.15" and "2.16".

With this change, everytime a new Git tag is pushed to the main
repository a new release is made. So we need to push tags before
publishing the repository to Hex.pm. This is due the release workflow
that have to build files before we use them.