Fix override of target by env vars on macOS (#33)
This is related to the download of precompiled NIF.

We now check the OS from the target that may be overriden by the
TARGET_OS env var. Before that we were using `:os.type/1` as reference
for normalization, but this won´t work if the target is different from
the OS returned by that function.

This also fix the normalization of vendor. It sets to "unknown" if some
env var changed the target but didn't specify a vendor.

This fix the target selection for people building for Nerves on macOS.