Add integrity check before extracting precompiled NIFs (#36)
This commit adds the integrity check of downloaded tar.gz files
containing the NIFs we published. The idea is avoid supply chain
attacks. If GitHub gets hacked we don't load unsafe files in our
machines.

The requirement is that before publishing the package to Hex,
we run the `mix rustler.download ModuleName --all` to generate
the checksum file for the native module.